requirements.txt
black
Django==4.1.7
django-cors-headers
djangorestframework==3.14.0
djangorestframework-camel-case
drf-spectacular
gunicorn
uvicorn
psycopg2-binary
django-allauth
dj-rest-auth
djangorestframework-simplejwt